+++ inc/CPU/ICPU.h
@@ -22,8 +22,8 @@ public:
     virtual uint8_t fetchByte() = 0;
     virtual uint16_t fetchWord() = 0;
     
-    virtual void writeByte(uint8_t address, uint8_t byte) = 0;
-    virtual void writeWord(uint8_t address, uint16_t word) = 0;
+    virtual void writeByte(uint16_t address, uint8_t byte) = 0;
+    virtual void writeWord(uint16_t address, uint16_t word) = 0;
     
     virtual void pushByte(uint8_t data) = 0;
     virtual uint8_t popByte() = 0;
